refactor: rename CLI package to Shell
- Renamed CLI.tsx to Shell.tsx and updated all references in the codebase. - Adjusted import paths in the relevant files to reflect the new package name. - Updated related test files and ensured they point to the correct module. - Verified that all functionalities are working correctly post-refactor. This change improves the clarity of the package's purpose and aligns with the overall project structure.
